Refactor received packet querier
