+++ crates/relayer-cosmos/src/impls/chain/ofa.rs
@@ -14,6 +14,7 @@ use ibc_relayer_components::chain::traits::components::counterparty_chain_id_que
 use ibc_relayer_components::chain::traits::components::create_client_message_builder::CreateClientMessageBuilder;
 use ibc_relayer_components::chain::traits::components::create_client_payload_builder::CreateClientPayloadBuilder;
 use ibc_relayer_components::chain::traits::components::message_sender::MessageSender;
+use ibc_relayer_components::chain::traits::components::received_packet_querier::ReceivedPacketQuerier;
 use ibc_relayer_components::chain::traits::components::write_ack_querier::WriteAckQuerier;
 use ibc_relayer_runtime::types::error::Error as TokioError;
 use ibc_relayer_runtime::types::log::logger::TracingLogger;
@@ -41,6 +42,7 @@ use crate::impls::chain::components::query_chain_id::QueryChainIdWithChainHandle
 use crate::impls::chain::components::query_chain_status::QueryChainStatusWithChainHandle;
 use crate::impls::chain::components::query_client_state::QueryCosmosClientStateFromChainHandle;
 use crate::impls::chain::components::query_consensus_state::QueryCosmosConsensusStateFromChainHandle;
+use crate::impls::chain::components::query_received_packet::QueryReceivedPacketWithChainHandle;
 use crate::impls::chain::components::query_write_ack_event::QueryWriteAckEventFromChainHandle;
 use crate::impls::chain::components::send_messages_as_tx::SendMessagesToTxContext;
 use crate::methods::channel::{
@@ -58,7 +60,7 @@ use crate::methods::event::{
 use crate::methods::packet::{
     build_ack_packet_message, build_ack_packet_payload, build_receive_packet_message,
     build_receive_packet_payload, build_timeout_unordered_packet_message,
-    build_timeout_unordered_packet_payload, query_is_packet_received,
+    build_timeout_unordered_packet_payload,
 };
 use crate::methods::unreceived_packet::{
     query_packet_commitments, query_send_packets_from_sequences, query_unreceived_packet_sequences,
@@ -564,7 +566,11 @@ where
         channel_id: &ChannelId,
         sequence: &Sequence,
     ) -> Result<bool, Error> {
-        query_is_packet_received(self, port_id, channel_id, sequence).await
+        <QueryReceivedPacketWithChainHandle as ReceivedPacketQuerier<
+            Self,
+            CosmosChain<Counterparty>,
+        >>::query_is_packet_received(self, port_id, channel_id, sequence)
+        .await
     }
 
     /// Query the sequences of the packets that the chain has committed to be
